// .PRD/erp-refactor/v3/types/payment-request.ts
import { z } from 'zod';

export const PaymentRequestSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  docstatus: z.number().default(0),
  payment_request_type: z.enum(['Outward', 'Inward']).default('Inward'),
  transaction_date: z.string().optional(),
  failed_reason: z.string().optional(),
  naming_series: z.enum(['ACC-PRQ-.YYYY.-']),
  company: z.string().optional(),
  mode_of_payment: z.string().optional(),
  party_type: z.string().optional(),
  party: z.string().optional(),
  party_name: z.string().optional(),
  reference_doctype: z.string().optional(),
  reference_name: z.string().optional(),
  grand_total: z.number(),
  currency: z.string().optional(),
  is_a_subscription: z.boolean().optional().default(false),
  outstanding_amount: z.number().optional(),
  party_account_currency: z.string().optional(),
  subscription_plans: z.array(z.unknown()).optional(),
  bank_account: z.string().optional(),
  bank: z.string().optional(),
  bank_account_no: z.string().optional(),
  account: z.string().optional(),
  iban: z.string().optional(),
  branch_code: z.string().optional(),
  swift_number: z.string().optional(),
  cost_center: z.string().optional(),
  project: z.string().optional(),
  print_format: z.string().optional(),
  email_to: z.string().optional(),
  subject: z.string().optional(),
  payment_gateway_account: z.string().optional(),
  status: z.enum(['Draft', 'Requested', 'Initiated', 'Partially Paid', 'Payment Ordered', 'Paid', 'Failed', 'Cancelled']).optional().default('Draft'),
  make_sales_invoice: z.boolean().optional().default(false),
  message: z.string().optional(),
  message_examples: z.string().optional(),
  mute_email: z.boolean().optional().default(false),
  payment_gateway: z.string().optional(),
  payment_account: z.string().optional(),
  payment_channel: z.enum(['Email', 'Phone', 'Other']).optional(),
  payment_order: z.string().optional(),
  amended_from: z.string().optional(),
  payment_url: z.string().max(500).optional(),
  phone_number: z.string().optional(),
});

export type PaymentRequest = z.infer<typeof PaymentRequestSchema>;

export const PaymentRequestInsertSchema = PaymentRequestSchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });
export type PaymentRequestInsert = z.infer<typeof PaymentRequestInsertSchema>;